Extract Admin interface from XXXAdmin
