+++ core/src/main/java/com/scalar/db/util/TableMetadataManager.java
@@ -3,7 +3,7 @@ package com.scalar.db.util;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import com.scalar.db.api.DistributedStorageAdmin;
+import com.scalar.db.api.Admin;
 import com.scalar.db.api.Operation;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
@@ -19,7 +19,7 @@ public class TableMetadataManager {
 
   private final LoadingCache<TableKey, Optional<TableMetadata>> tableMetadataCache;
 
-  public TableMetadataManager(DistributedStorageAdmin admin, long cacheExpirationTimeSecs) {
+  public TableMetadataManager(Admin admin, long cacheExpirationTimeSecs) {
     CacheBuilder<Object, Object> builder = CacheBuilder.newBuilder();
     if (cacheExpirationTimeSecs > 0) {
       builder.expireAfterWrite(cacheExpirationTimeSecs, TimeUnit.SECONDS);
